tests/wpt/tests/html/canvas/offscreen/text/2d.text.fontVariantCaps2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.fontVariantCaps2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.fontVariantCaps2</h1>
<p class="desc">Testing small caps setting in fontVariant</p>


<script>
var t = async_test("Testing small caps setting in fontVariant");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.font = "small-caps 32px serif";
  // "mismatch" test, to verify that small-caps does change the rendering.
  smallCaps_len = ctx.measureText("Hello World").width;

  ctx.font = "32px serif";
  normalCaps_len = ctx.measureText("Hello World").width;
  _assert(smallCaps_len != normalCaps_len, "smallCaps_len != normalCaps_len");
  t.done();

});
</script>
